:art: refactor: 이름 변경 및 코드 위치 변경
