+++ app/src/main/java/com/example/swap_it/data/datasource/local/model/bottomsheet/CategoryOption.kt
@@ -1,0 +1,14 @@
+package com.example.swap_it.data.datasource.local.model.bottomsheet
+
+enum class CategoryOption(val categoryName: String) {
+    CAR("자동차"),
+    CLOTH("옷"),
+    MOBILE("모바일 기기"),
+    CUP("컵"),
+    BOOK("책"),
+    ELECTRONIC("가전기기"),
+    FASHION("패션"),
+    FOOD("음식"),
+    ETC("기타"),
+    SNACK("과자"),
+}
