+++ app/src/main/java/com/example/swap_it/data/datasource/local/model/bottomsheet/SortOption.kt
@@ -1,0 +1,8 @@
+package com.example.swap_it.data.datasource.local.model.bottomsheet
+
+enum class SortOption(val option: String) {
+    POPULAR("인기순"),
+    RECENT("최신순"),
+    PRICE_HIGH("가격 높은 순"),
+    PRICE_ROW("가격 낮은 순"),
+}
