+++ app/src/main/java/com/example/swap_it/data/datasource/local/model/search/CurrentSearch.kt
@@ -1,0 +1,14 @@
+package com.example.swap_it.data.datasource.local.model.search
+
+enum class CurrentSearch(val categoryName: String) {
+        CAR("미미"),
+        CLOTH("부족해"),
+        MOBILE("가방"),
+        CUP("나같은"),
+        BOOK("없다는"),
+        ELECTRONIC("걸"),
+        FASHION("아니"),
+        FOOD("시간"),
+        ETC("이"),
+        SNACK("흐른"),
+    }
