+++ app/src/main/java/com/example/swap_it/data/datasource/local/model/search/PopularSearch.kt
@@ -1,0 +1,14 @@
+package com.example.swap_it.data.datasource.local.model.search
+
+enum class PopularSearch(val categoryName: String) {
+        CAR("참기 힘들었어"),
+        CLOTH("내 맘을 보여주면"),
+        MOBILE("변할지 난 궁금해"),
+        CUP("이 바보같은 모습을"),
+        BOOK("참 많이 부족하지만"),
+        ELECTRONIC("나 같은 남잔 없다는 걸 아니"),
+        FASHION("많이 부족하겠지만"),
+        FOOD("세월이 흘러가도"),
+        ETC("너의 작은 실수라도"),
+        SNACK("많이"),
+    }
