+++ app/src/main/java/com/example/swap_it/ui/component/Cards.kt
@@ -10,7 +10,9 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.MaterialTheme.colors
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Send
 import androidx.compose.material.icons.filled.Settings
@@ -23,14 +25,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.painter.ColorPainter
 import androidx.compose.ui.graphics.vector.rememberVectorPainter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import coil3.compose.AsyncImage
+import coil.compose.AsyncImage
 import com.example.swap_it.R
-import com.example.swap_it.ui.alert.AlertCardData
 import com.example.swap_it.ui.theme.BackgroundColor
 import com.example.swap_it.ui.theme.Gray3
 import com.example.swap_it.ui.theme.Gray4
@@ -40,7 +42,7 @@ import com.example.swap_it.ui.theme.Typography
 import com.example.swap_it.ui.theme.White
 
 
-data class ListCardData(
+data class ShoppingCardData(
     val imageUri: String,
     val category: String,
     val viewCount: String,
@@ -51,86 +53,106 @@ data class ListCardData(
     val onClick: () -> Unit = {}
 )
 
+data class AlertCardData(
+    val message: String,
+    val date: String,
+    val icon: Int,
+    val onClick: () -> Unit = {}
+)
+
+data class SwapCardData(
+    val imageUri: String,
+    val category: String,
+    val viewCount: String,
+    val region: String,
+    val time: String,
+    val price: String,
+    val title: String,
+    val onClick: () -> Unit = {}
+)
+
+
 class Cards {
     @OptIn(ExperimentalMaterial3Api::class)
     @Composable
-    fun ProductListCard(listCardData: ListCardData) {
+    fun ShoppingCard(shoppingCardData: ShoppingCardData) {
         Card(
-            modifier = Modifier.padding(Paddings.xlarge, Paddings.smallMedium),
+            modifier = Modifier.fillMaxWidth().padding(Paddings.xlarge, Paddings.smallMedium),
             colors = CardDefaults.cardColors(
                 containerColor = White
             ),
-            shape = RoundedCornerShape(8.dp),
-            onClick = listCardData.onClick
+            shape = RoundedCornerShape(20.dp),
+            onClick = shoppingCardData.onClick
         ) {
-            Row {
+            Row (verticalAlignment = Alignment.CenterVertically){
+                Spacer(modifier = Modifier.size(Paddings.large))
                 AsyncImage(
-                    model = listCardData.imageUri,
+                    model = shoppingCardData.imageUri,
                     contentDescription = "상품 대표 이미지",
                     modifier = Modifier
-                        .size(100.dp)
-                        .clip(RoundedCornerShape(4.dp)),
+                        .size(86.dp)
+                        .clip(RoundedCornerShape(12.dp)),
                     placeholder = ColorPainter(Primary),
                     fallback = rememberVectorPainter(Icons.Default.Send),
                     error = rememberVectorPainter(Icons.Default.Settings),
                 )
-                Column(modifier = Modifier.padding(Paddings.medium)) {
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text(
-                            text = listCardData.category,
-                            style = Typography.labelLarge,
-                            color = Gray4
-                        )
-                        Text(" | ", color = Gray4)
-                        Text(
-                            text = listCardData.region,
-                            style = Typography.labelLarge,
-                            color = Gray4
-                        )
-                        Text(" | ", color = Gray4)
-                        Text(
-                            text = listCardData.time,
-                            style = Typography.labelLarge,
-                            color = Gray4
-                        )
-                    }
-                    Spacer(modifier = Modifier.height(Paddings.medium))
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text(
-                            text = listCardData.title,
-                            style = Typography.titleMedium
-                        )
-                    }
-                    Spacer(modifier = Modifier.height(Paddings.medium))
-                    Row(verticalAlignment = Alignment.CenterVertically) {
+                Column(modifier = Modifier.padding(Paddings.large, Paddings.none, Paddings.none, Paddings.none)) {
+                    Text(
+                        modifier = Modifier.padding(
+                            Paddings.none,
+                            Paddings.largeExtra,
+                            Paddings.none,
+                            Paddings.xsmall
+                        ),
+                        text = "${shoppingCardData.category} | ${shoppingCardData.region} | ${shoppingCardData.time}",
+                        style = Typography.labelLarge,
+                        color = Gray4
+                    )
+                    Text(
+                        modifier = Modifier.padding(
+                            Paddings.none,
+                            Paddings.none,
+                            Paddings.none,
+                            Paddings.small
+                        ),
+                        text = shoppingCardData.title,
+                        style = Typography.titleMedium,
+                        maxLines = 1,
+                    )
+                    Row(
+                        modifier = Modifier.padding(
+                            Paddings.none,
+                            Paddings.none,
+                            Paddings.extra,
+                            Paddings.largeExtra
+                        ), verticalAlignment = Alignment.CenterVertically
+                    ) {
                         Row(
                             modifier = Modifier.weight(1f),
                             verticalAlignment = Alignment.CenterVertically
                         ) {
                             Text(
                                 text = "예상 ",
-                                style = Typography.titleMedium,
+                                style = Typography.bodyMedium,
                                 color = Gray3
                             )
                             Text(
-                                text = listCardData.price,
+                                text = shoppingCardData.price,
                                 style = Typography.titleMedium,
                             )
                             Text(
                                 text = "원",
-                                style = Typography.titleMedium,
+                                style = Typography.bodyMedium,
                                 color = Gray3
                             )
                         }
                         Image(
                             painter = painterResource(id = R.drawable.ic_show),
                             contentDescription = "조회 수",
-                            modifier = Modifier.height(30.dp),
-                            colorFilter = androidx.compose.ui.graphics.ColorFilter.tint(Gray4)
+                            colorFilter = ColorFilter.tint(Gray4)
                         )
-                        Spacer(modifier = Modifier.size(Paddings.small))
                         Text(
-                            text = listCardData.viewCount,
+                            text = shoppingCardData.viewCount,
                             style = Typography.labelLarge,
                             color = Gray4
                         )
@@ -144,7 +166,7 @@ class Cards {
 
 
     @Composable
-    fun AlertListCard(alertCardData: AlertCardData) {
+    fun AlertCard(alertCardData: AlertCardData) {
         @OptIn(ExperimentalMaterial3Api::class)
         Card(
             modifier = Modifier.padding(Paddings.xlarge, Paddings.medium),
@@ -195,20 +217,113 @@ class Cards {
         }
     }
 
+    @OptIn(ExperimentalMaterial3Api::class)
+    @Composable
+    fun SwapCard(swapCardData: SwapCardData) {
+        Card(
+            modifier = Modifier
+                .width(162.dp)
+                .padding(Paddings.smallMedium),
+            colors = CardDefaults.cardColors(
+                containerColor = White
+            ),
+            shape = RoundedCornerShape(20.dp),
+            onClick = swapCardData.onClick
+        ) {
+            Column {
+                AsyncImage(
+                    model = swapCardData.imageUri,
+                    contentDescription = "상품 대표 이미지",
+                    modifier = Modifier
+                        .size(150.dp),
+                    placeholder = ColorPainter(Primary),
+                    fallback = rememberVectorPainter(Icons.Default.Send),
+                    error = rememberVectorPainter(Icons.Default.Settings),
+                )
+                Column(modifier = Modifier.padding(Paddings.large)) {
+                    Text(
+                        text = "${swapCardData.category} | ${swapCardData.region}",
+                        style = Typography.labelLarge,
+                        color = Gray4
+                    )
+                    Text(
+                        text = swapCardData.title,
+                        style = Typography.titleMedium,
+                        maxLines = 1,
+                        modifier = Modifier.padding(Paddings.none,Paddings.none,Paddings.none,Paddings.small)
+                    )
+
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically,
+                        modifier = Modifier.padding(Paddings.none,Paddings.none,Paddings.none,Paddings.large)
+                    ) {
+                        Text(
+                            text = "예상 ",
+                            style = Typography.titleMedium,
+                            color = Gray3
+                        )
+                        Text(
+                            text = swapCardData.price,
+                            style = Typography.titleMedium,
+                        )
+                        Text(
+                            text = "원",
+                            style = Typography.titleMedium,
+                            color = Gray3
+                        )
+                    }
+                    Row(verticalAlignment = Alignment.CenterVertically) {
+                        Text(
+                            modifier = Modifier.weight(1f),
+                            text = swapCardData.time,
+                            style = Typography.labelLarge,
+                            color = Gray4
+                        )
+                        Image(
+                            painter = painterResource(id = R.drawable.ic_show),
+                            contentDescription = "조회 수",
+                            modifier = Modifier.height(30.dp),
+                            colorFilter = ColorFilter.tint(Gray4)
+                        )
+                        Spacer(modifier = Modifier.size(Paddings.small))
+                        Text(
+                            text = swapCardData.viewCount,
+                            style = Typography.labelLarge,
+                            color = Gray4
+                        )
+                    }
+
+                }
+
+            }
+
+        }
+    }
+
+
     companion object {
-        val productCardData = ListCardData(
-            imageUri = "https://velog.velcdn.com/images/deepblue/post/f657fa91-f059-4e8d-81c3-c7c7776e0d9f/image.png",
+        val productCardData = ShoppingCardData(
+            imageUri = "https://static.nike.com/a/images/c_limit,w_592,f_auto/t_product_v1/d0382875-94af-4a2f-a788-6d2ce0657de2/W+AIR+MAX+DN8.png",
             category = "가방",
             viewCount = "100",
             region = "강서구",
             time = "1일전",
             price = "10000",
-            title = "가방 팔아요~~~",
+            title = "나이키 운동화",
         )
         val alertCardData = AlertCardData(
             message = "스왑 요청이 들어왔어요",
             date = "2월 19일 13:22",
             icon = R.drawable.ic_show
+        )
+        val swapCardData = SwapCardData(
+            imageUri = "https://static.nike.com/a/images/c_limit,w_592,f_auto/t_product_v1/d0382875-94af-4a2f-a788-6d2ce0657de2/W+AIR+MAX+DN8.png",
+            category = "가방",
+            viewCount = "100",
+            region = "강서구",
+            time = "1일전",
+            price = "10000",
+            title = "나이키 운동화",
         )
     }
 }
@@ -216,8 +331,24 @@ class Cards {
 
 @Preview(showBackground = true)
 @Composable
-fun DefaultPreview() {
-    Cards().ProductListCard(
+fun ShoppingCardPreview() {
+    Cards().ShoppingCard(
         Cards.productCardData
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun AlertCardPreview() {
+    Cards().AlertCard(
+        Cards.alertCardData
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun SwapCardPreview() {
+    Cards().SwapCard(
+        Cards.swapCardData
     )
 }
