+++ app/src/main/java/com/example/swap_it/ui/shopping/ModalBottomSheetContent.kt
@@ -21,6 +21,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import com.example.swap_it.R
+import com.example.swap_it.data.datasource.local.model.bottomsheet.CategoryOption
+import com.example.swap_it.data.datasource.local.model.bottomsheet.SortOption
 import com.example.swap_it.ui.component.CategoryButton
 import com.example.swap_it.ui.theme.Gray5
 import com.example.swap_it.ui.theme.Paddings
@@ -42,12 +44,7 @@ fun ModalBottomSheetContent() {
 
 
 
-enum class SortOption(val option: String) {
-    POPULAR("인기순"),
-    RECENT("최신순"),
-    PRICE_HIGH("가격 높은 순"),
-    PRICE_ROW("가격 낮은 순"),
-}
+
 
 @Composable
 fun SortButtons(
@@ -93,18 +90,7 @@ fun SortButton(
 
 }
 
-enum class CategoryOption(val categoryName: String) {
-    CAR("자동차"),
-    CLOTH("옷"),
-    MOBILE("모바일 기기"),
-    CUP("컵"),
-    BOOK("책"),
-    ELECTRONIC("가전기기"),
-    FASHION("패션"),
-    FOOD("음식"),
-    ETC("기타"),
-    SNACK("과자"),
-}
+
 
 @OptIn(ExperimentalLayoutApi::class)
 @Composable
