+++ app/src/main/java/com/example/swap_it/ui/swap/SwapScreen.kt
@@ -1,0 +1,44 @@
+package com.example.swap_it.ui.swap
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.NavHostController
+import com.example.swap_it.ui.component.AppBar
+import com.example.swap_it.ui.component.BottomNavigationBar
+import com.example.swap_it.ui.theme.BackgroundColor
+
+
+@Composable
+fun SwapScreen(modifier: Modifier, navController: NavHostController) {
+    Scaffold(
+        topBar = {
+            AppBar(navController = navController)
+        },
+        bottomBar = {
+            BottomNavigationBar(navController)
+        },
+    ) { contentPadding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(contentPadding)
+                .background(BackgroundColor)
+        ) {
+            RequestedSwapSection(modifier, navController)
+            RequestSwapSection(modifier, navController)
+        }
+    }
+}
+
+@Composable
+@Preview(showBackground = true)
+fun SwapScreenPreview() {
+    SwapScreen(Modifier, navController = NavHostController(context = LocalContext.current))
+}
